refactor: 토큰 관련 DTO를 비즈니스 목적과 시스템 목적으로 분류
